matchcenterai/pyproject.toml
[build-system]
requires = [ "setuptools", "setuptools-scm" ]
build-backend = "setuptools.build_meta"

[project]
name = "matchcenterai"
version = "0.1.0"
authors = [{ name = "Koen Benne", email = "koen.benne@iodigital.com" }]

dependencies = [
  "blogger",
  "summarizer",
  "PyAudio",
  "pyyaml",
  "fastapi",
  "uvicorn",
]

[project.scripts]
matchcenterai = "matchcenterai:main"
